handle initial install: when file is missing

+++ Rakefile
@@ -39,7 +39,9 @@ class DevUtilsHelper
       util[:binaries].each do |bin|
         src_file  = File.join($proj_root,util[:name],'bin',bin)
         dest_file = File.join(bin_dir,bin)
-        FileUtils.rm dest_file
+        if File.exist? dest_file
+          FileUtils.rm dest_file
+        end
         unless File.exist? dest_file
           FileUtils.ln_s src_file, dest_file
         end
